+++ tests/test_cli.py
@@ -9,11 +9,13 @@ test_depoy_nonempty_file -- test embedding a deployment password in a nonempty f
 test_environment_variable_empty_file -- test embedding an environment variable in an empty file
 test_environment_variable_nonempty_file -- test embedding an environment variable in a nonempty file
 """
+import base64
+from collections import OrderedDict
 
 from click.testing import CliRunner
-import yaml
 
 from travis.cli import cli
+from travis.orderer import ordered_load, ordered_dump
 
 
 def test_password_output():
@@ -22,6 +24,8 @@ def test_password_output():
     result = runner.invoke(cli, ['mandeep', 'Travis-Encrypt'],
                            'SUPER_SECURE_PASSWORD')
     assert not result.exception
+    assert 'Password: \n\nPlease add the following to your .travis.yml:\nsecure:' in result.output
+    assert base64.b64decode(result.output.split()[-1])
 
 
 def test_password_empty_file():
@@ -30,11 +34,19 @@ def test_password_empty_file():
     with runner.isolated_filesystem():
         initial_data = {'language': 'python'}
         with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
+            ordered_dump(initial_data, file)
 
         result = runner.invoke(cli, ['mandeep', 'Travis-Encrypt', 'file.yml'],
                                'SUPER_SECURE_PASSWORD')
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+            assert 'password' in config
+            assert 'secure' in config['password']
+            assert config['language'] == 'python'
+            assert base64.b64decode(config['password']['secure'])
 
 
 def test_password_nonempty_file():
@@ -43,14 +55,24 @@ def test_password_nonempty_file():
     The YAML file includes information that needs to be overwritten."""
     runner = CliRunner()
     with runner.isolated_filesystem():
-        initial_data = {'language': 'python', 'dist': 'trusty',
-                        'password': {'secure': 'SUPER_INSECURE_PASSWORD'}}
+
+        initial_data = OrderedDict([('language', 'python'), ('dist', 'trusty'),
+                                    ('password', {'secure': 'SUPER_INSECURE_PASSWORD'})])
+
         with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
+            ordered_dump(initial_data, file)
 
         result = runner.invoke(cli, ['mandeep', 'Travis-Encrypt', 'file.yml'],
                                'SUPER_SECURE_PASSWORD')
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+        assert config['language'] == 'python'
+        assert config['dist'] == 'trusty'
+        assert base64.b64decode(config['password']['secure'])
+        assert ['language', 'dist', 'password'] == [key for key in config.keys()]
 
 
 def test_deploy_empty_file():
@@ -59,11 +81,17 @@ def test_deploy_empty_file():
     with runner.isolated_filesystem():
         initial_data = {'language': 'python'}
         with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
+            ordered_dump(initial_data, file)
 
         result = runner.invoke(cli, ['--deploy', 'mandeep', 'Travis-Encrypt', 'file.yml'],
                                'SUPER_SECURE_PASSWORD')
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+        assert config['language'] == 'python'
+        assert base64.b64decode(config['deploy']['password']['secure'])
 
 
 def test_deploy_nonempty_file():
@@ -72,27 +100,48 @@ def test_deploy_nonempty_file():
     The YAML file includes information that needs to be overwritten."""
     runner = CliRunner()
     with runner.isolated_filesystem():
-        initial_data = {'language': 'python', 'dist': 'trusty',
-                        'deploy': {'password': {'secure': 'SUPER_INSECURE_PASSWORD'}}}
+
+        initial_data = OrderedDict([('language', 'python'), ('dist', 'trusty'),
+                                    ('deploy', {'password': {'secure': 'SUPER_INSECURE_PASSWORD'}})])
+
         with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
+            ordered_dump(initial_data, file)
 
         result = runner.invoke(cli, ['--deploy', 'mandeep', 'Travis-Encrypt', 'file.yml'],
                                'SUPER_SECURE_PASSWORD')
+
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+        assert config['language'] == 'python'
+        assert config['dist'] == 'trusty'
+        assert base64.b64decode(config['deploy']['password']['secure'])
+        assert ['language', 'dist', 'deploy'] == [key for key in config.keys()]
 
 
 def test_environment_variable_empty_file():
     """Test the encrypt module's CLI function with the --env flag and an empty YAML file."""
     runner = CliRunner()
     with runner.isolated_filesystem():
+
         initial_data = {'language': 'python'}
+
         with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
+            ordered_dump(initial_data, file)
 
         result = runner.invoke(cli, ['--env', 'mandeep', 'Travis-Encrypt', 'file.yml'],
                                'API_KEY=SUPER_SECURE_KEY')
+
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+        assert config['language'] == 'python'
+        assert base64.b64decode(config['env']['global']['secure'])
+        assert ['language', 'env'] == [key for key in config.keys()]
 
 
 def test_environment_variable_nonempty_file():
@@ -101,11 +150,22 @@ def test_environment_variable_nonempty_file():
     The YAML file includes information that needs to be overwritten."""
     runner = CliRunner()
     with runner.isolated_filesystem():
-        initial_data = {'language': 'python', 'dist': 'trusty',
-                        'env': {'global': {'secure': 'API_KEY=SUPER_INSECURE_KEY'}}}
-        with open('file.yml', 'w') as file:
-            yaml.safe_dump(initial_data, file, default_flow_style=True)
 
-        result = runner.invoke(cli, ['mandeep', 'Travis-Encrypt', 'file.yml'],
+        initial_data = OrderedDict([('language', 'python'), ('dist', 'trusty'),
+                                    ('env', {'global': {'secure': 'API_KEY="SUPER_INSECURE_KEY"'}})])
+
+        with open('file.yml', 'w') as file:
+            ordered_dump(initial_data, file)
+
+        result = runner.invoke(cli, ['--env', 'mandeep', 'Travis-Encrypt', 'file.yml'],
                                'SUPER_SECURE_API_KEY')
+
         assert not result.exception
+
+        with open('file.yml') as file:
+            config = ordered_load(file)
+
+        assert config['language'] == 'python'
+        assert config['dist'] == 'trusty'
+        assert base64.b64decode(config['env']['global']['secure'])
+        assert ['language', 'dist', 'env'] == [key for key in config.keys()]